fix(#1014): move sync allergens logic into hook to satisfy ESLint import rules
